+++ backend/app/models/permission.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime, timezone
 
-from sqlalchemy import BigInteger, DateTime, String
+from sqlalchemy import Integer, DateTime, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database.db import Base
@@ -11,7 +11,7 @@ from app.database.db import Base
 class Permission(Base):
     __tablename__ = "permissions"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     code: Mapped[str] = mapped_column(String(150), unique=True, nullable=False)
     description: Mapped[str | None] = mapped_column(String(255), nullable=True)
     module: Mapped[str | None] = mapped_column(String(100), nullable=True)
